+++ lib/compat/wordpress-6.5/interactivity-api/interactivity-api.php
@@ -4,6 +4,7 @@
  *
  * @package WordPress
  * @subpackage Interactivity API
+ * @since 6.5.0
  */
 
 if ( ! function_exists( 'wp_interactivity_process_directives_of_interactive_blocks' ) ) {
@@ -20,7 +21,7 @@ if ( ! function_exists( 'wp_interactivity_process_directives_of_interactive_bloc
 	 * @param array $parsed_block The parsed block.
 	 * @return array The same parsed block.
 	 */
-	function wp_interactivity_process_directives_of_interactive_blocks( $parsed_block ) {
+	function wp_interactivity_process_directives_of_interactive_blocks( array $parsed_block ): array {
 		static $root_interactive_block = null;
 
 		/*
@@ -31,38 +32,42 @@ if ( ! function_exists( 'wp_interactivity_process_directives_of_interactive_bloc
 			$block_name = $parsed_block['blockName'];
 			$block_type = WP_Block_Type_Registry::get_instance()->get_registered( $block_name );
 
-			if ( isset( $block_name ) && isset( $block_type->supports['interactivity'] ) && $block_type->supports['interactivity'] ) {
+			if (
+			isset( $block_name ) &&
+			( ( isset( $block_type->supports['interactivity'] ) && true === $block_type->supports['interactivity'] ) ||
+			( isset( $block_type->supports['interactivity']['interactive'] ) && true === $block_type->supports['interactivity']['interactive'] ) )
+			) {
 				// Annotates the root interactive block for processing.
-				$root_interactive_block = array( $block_name, md5( serialize( $parsed_block ) ) );
+				$root_interactive_block = array( $block_name, $parsed_block );
 
 				/*
 				 * Adds a filter to process the root interactive block once it has
 				 * finished rendering.
 				 */
-				$process_interactive_blocks = static function ( $content, $parsed_block ) use ( &$root_interactive_block, &$process_interactive_blocks ) {
+				$process_interactive_blocks = static function ( string $content, array $parsed_block ) use ( &$root_interactive_block, &$process_interactive_blocks ): string {
 					// Checks whether the current block is the root interactive block.
-					list($root_block_name, $root_block_md5) = $root_interactive_block;
-					if ( $root_block_name === $parsed_block['blockName'] && md5( serialize( $parsed_block ) ) === $root_block_md5 ) {
+					list($root_block_name, $root_parsed_block) = $root_interactive_block;
+					if ( $root_block_name === $parsed_block['blockName'] && $parsed_block === $root_parsed_block ) {
 						// The root interactive blocks has finished rendering, process it.
 						$content = wp_interactivity_process_directives( $content );
 						// Removes the filter and reset the root interactive block.
-						remove_filter( 'render_block', $process_interactive_blocks );
+						remove_filter( 'render_block_' . $parsed_block['blockName'], $process_interactive_blocks );
 						$root_interactive_block = null;
 					}
 					return $content;
 				};
 
 				/*
-				 * Uses a priority of 20 to ensure that other filters can add additional
-				 * directives before the processing starts.
-				 */
-				add_filter( 'render_block', $process_interactive_blocks, 20, 2 );
+				* Uses a priority of 20 to ensure that other filters can add additional
+				* directives before the processing starts.
+				*/
+				add_filter( 'render_block_' . $block_name, $process_interactive_blocks, 20, 2 );
 			}
 		}
 
 		return $parsed_block;
 	}
-	add_filter( 'render_block_data', 'wp_interactivity_process_directives_of_interactive_blocks', 10, 1 );
+	add_filter( 'render_block_data', 'wp_interactivity_process_directives_of_interactive_blocks' );
 }
 
 if ( ! function_exists( 'wp_interactivity' ) ) {
@@ -70,22 +75,23 @@ if ( ! function_exists( 'wp_interactivity' ) ) {
 	 * Retrieves the main WP_Interactivity_API instance.
 	 *
 	 * It provides access to the WP_Interactivity_API instance, creating one if it
-	 * doesn't exist yet. It also registers the hooks and necessary script
-	 * modules.
+	 * doesn't exist yet.
+	 *
+	 * @global WP_Interactivity_API $wp_interactivity
 	 *
 	 * @since 6.5.0
 	 *
 	 * @return WP_Interactivity_API The main WP_Interactivity_API instance.
 	 */
-	function wp_interactivity() {
-		static $instance = null;
-		if ( is_null( $instance ) ) {
-			$instance = new WP_Interactivity_API();
-			$instance->add_hooks();
-			$instance->register_script_modules();
+	function wp_interactivity(): WP_Interactivity_API {
+		global $wp_interactivity;
+		if ( ! ( $wp_interactivity instanceof WP_Interactivity_API ) ) {
+			$wp_interactivity = new WP_Interactivity_API();
 		}
-		return $instance;
+		return $wp_interactivity;
 	}
+
+	wp_interactivity()->add_hooks();
 }
 
 if ( ! function_exists( 'wp_interactivity_process_directives' ) ) {
@@ -98,7 +104,7 @@ if ( ! function_exists( 'wp_interactivity_process_directives' ) ) {
 	 * @param string $html The HTML content to process.
 	 * @return string The processed HTML content. It returns the original content when the HTML contains unbalanced tags.
 	 */
-	function wp_interactivity_process_directives( $html ) {
+	function wp_interactivity_process_directives( string $html ): string {
 		return wp_interactivity()->process_directives( $html );
 	}
 }
@@ -116,9 +122,10 @@ if ( ! function_exists( 'wp_interactivity_state' ) ) {
 	 * @param string $store_namespace The unique store namespace identifier.
 	 * @param array  $state           Optional. The array that will be merged with the existing state for the specified
 	 *                                store namespace.
-	 * @return array The current state for the specified store namespace.
+	 * @return array The state for the specified store namespace. This will be the updated state if a $state argument was
+	 *               provided.
 	 */
-	function wp_interactivity_state( $store_namespace, $state = null ) {
+	function wp_interactivity_state( string $store_namespace, array $state = array() ): array {
 		return wp_interactivity()->state( $store_namespace, $state );
 	}
 }
@@ -136,9 +143,38 @@ if ( ! function_exists( 'wp_interactivity_config' ) ) {
 	 * @param string $store_namespace The unique store namespace identifier.
 	 * @param array  $config          Optional. The array that will be merged with the existing configuration for the
 	 *                                specified store namespace.
-	 * @return array The current configuration for the specified store namespace.
+	 * @return array The configuration for the specified store namespace. This will be the updated configuration if a
+	 *               $config argument was provided.
 	 */
-	function wp_interactivity_config( $store_namespace, $initial_state = null ) {
-		return wp_interactivity()->config( $store_namespace, $initial_state );
+	function wp_interactivity_config( string $store_namespace, array $config = array() ): array {
+		return wp_interactivity()->config( $store_namespace, $config );
+	}
+}
+
+if ( ! function_exists( 'data_wp_context' ) ) {
+	/**
+	 * Generates a `data-wp-context` directive attribute by encoding a context
+	 * array.
+	 *
+	 * This helper function simplifies the creation of `data-wp-context` directives
+	 * by providing a way to pass an array of data, which encodes into a JSON string
+	 * safe for direct use as a HTML attribute value.
+	 *
+	 * Example:
+	 *
+	 *     <div <?php echo data_wp_context( array( 'isOpen' => true, 'count' => 0 ) ); ?>>
+	 *
+	 * @since 6.5.0
+	 *
+	 * @param array  $context         The array of context data to encode.
+	 * @param string $store_namespace Optional. The unique store namespace identifier.
+	 * @return string A complete `data-wp-context` directive with a JSON encoded value representing the context array and
+	 *                the store namespace if specified.
+	 */
+	function data_wp_context( array $context, string $store_namespace = '' ): string {
+		return 'data-wp-context=\'' .
+		( $store_namespace ? $store_namespace . '::' : '' ) .
+		( empty( $context ) ? '{}' : wp_json_encode( $context, JSON_HEX_TAG | JSON_HEX_APOS | JSON_HEX_QUOT | JSON_HEX_AMP ) ) .
+		'\'';
 	}
 }